langchain_project/requirements.txt
pymongo
streamlit
langchain
python-dotenv
openai
bson